Layouts, Logo, AppIcon and other minor fixes done. Application is ready ship.
