+++ Scoreboard - Gully cricket/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,6 +1,15 @@
 {
   "colors" : [
     {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0xFF",
+          "green" : "0x7F",
+          "red" : "0x00"
+        }
+      },
       "idiom" : "universal"
     }
   ],
