--- Scoreboard - Gully cricket/Assets.xcassets/LaunchScreen.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "img3.wallspic.com-ball_game-ball-factory-plant-team_sport-3200x2142@1x.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "img3.wallspic.com-ball_game-ball-factory-plant-team_sport-3200x2142@2x.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "filename" : "img3.wallspic.com-ball_game-ball-factory-plant-team_sport-3200x2142@3x.png",
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
