+++ Scoreboard - Gully cricket/Scoreboard.swift
@@ -5,8 +5,10 @@ import SwiftUI
 @main
 struct  Scoreboard:App {
     
-    @StateObject var matchvariables = MatchVariables.matchVariables;
+    @StateObject var matchvariables = MatchVariables.matchVariables
+    
     let migrator = Migrator()
+    
     var body: some Scene {
         WindowGroup {
             ContentView()
